feat: Prevent install of existing font, and wezterm download.

+++ setup/install-wezterm.ps1
@@ -59,13 +59,18 @@ function Install-WezTermWindows {
 }
 
 function Install-WezTermLinux {
-    Write-Host "Installing WezTerm on Linux..."
-
-    curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o /etc/apt/keyrings/wezterm-fury.gpg
-    echo 'deb [signed-by=/etc/apt/keyrings/wezterm-fury.gpg] https://apt.fury.io/wez/ * *' | sudo tee /etc/apt/sources.list.d/wezterm.list
-
-    sudo apt update
-    sudo apt install -y wezterm-nightly
+    if (Test-CommandExists $appCmd) {
+        Write-Host "$appCmd command found."
+        Write-Host "Updating WezTerm on Linux..."
+        sudo apt install --only-upgrade -y wezterm-nightly
+    }
+    else {
+        Write-Host "$appCmd command not found."
+        Write-Host "Installing WezTerm on Linux..."
+        curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o /etc/apt/keyrings/wezterm-fury.gpg
+        echo 'deb [signed-by=/etc/apt/keyrings/wezterm-fury.gpg] https://apt.fury.io/wez/ * *' | sudo tee /etc/apt/sources.list.d/wezterm.list
+        sudo apt install -y wezterm-nightly
+    }
 }
 
 # run
